packages/api-python/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
supabase>=2.3.0
asyncpg>=0.29.0

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Payments
stripe>=7.0.0

# Rate limiting
slowapi>=0.1.9

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Production
gunicorn>=21.2.0
